fix: widen histogram bins to u32 to prevent overflow
With MAX_LEN=32768 bytes where both nibbles are identical (e.g. 0x00),
a single bin receives 2*32768=65536 increments, overflowing u16.

+++ dudero.h
@@ -17,7 +17,7 @@ typedef enum {
 } dudero_ret_t;
 
 typedef struct {
-    uint16_t hist[16]; // histogram bins, count up to 2^16 = 65,536
+    uint32_t hist[16]; // histogram bins
     size_t hist_samples; // total number of samples processed
 } dudero_ctx_t;
